--- samples/snippets/storage_transfer_manager.py
@@ -1,184 +1,0 @@
-# Copyright 2022 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the 'License');
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-
-def upload_many_blobs_with_transfer_manager(
-    bucket_name, filenames, source_directory="", threads=4
-):
-    """Upload every file in a list to a bucket, concurrently in a thread pool.
-
-    Each blob name is derived from the filename, not including the
-    `source_directory` parameter. For complete control of the blob name for each
-    file (and other aspects of individual blob metadata), use
-    transfer_manager.upload_many() instead.
-    """
-
-    # The ID of your GCS bucket
-    # bucket_name = "your-bucket-name"
-
-    # A list (or other iterable) of filenames to upload.
-    # filenames = ["file_1.txt", "file_2.txt"]
-
-    # The directory on your computer that is the root of all of the files in the
-    # list of filenames. This string is prepended (with os.path.join()) to each
-    # filename to get the full path to the file. Relative paths and absolute
-    # paths are both accepted. This string is not included in the name of the
-    # uploaded blob; it is only used to find the source files. An empty string
-    # means "the current working directory". Note that this parameter allows
-    # directory traversal (e.g. "/", "../") and is not intended for unsanitized
-    # end user input.
-    # source_directory=""
-
-    # The number of threads to use for the operation. The performance impact of
-    # this value depends on the use case, but generally, smaller files benefit
-    # from more threads and larger files don't benefit from more threads. Too
-    # many threads can slow operations, especially with large files, due to
-    # contention over the Python GIL.
-    # threads=4
-
-    from google.cloud.storage import Client, transfer_manager
-
-    storage_client = Client()
-    bucket = storage_client.bucket(bucket_name)
-
-    results = transfer_manager.upload_many_from_filenames(
-        bucket, filenames, source_directory=source_directory, threads=threads
-    )
-
-    for name, result in zip(filenames, results):
-        # The results list is either `None` or an exception for each filename in
-        # the input list, in order.
-
-        if isinstance(result, Exception):
-            print("Failed to upload {} due to exception: {}".format(name, result))
-        else:
-            print("Uploaded {} to {}.".format(name, bucket.name))
-
-
-def upload_directory_with_transfer_manager(bucket_name, source_directory, threads=4):
-    """Upload every file in a directory, including all files in subdirectories.
-
-    Each blob name is derived from the filename, not including the `directory`
-    parameter itself. For complete control of the blob name for each file (and
-    other aspects of individual blob metadata), use
-    transfer_manager.upload_many() instead.
-    """
-
-    # The ID of your GCS bucket
-    # bucket_name = "your-bucket-name"
-
-    # The directory on your computer to upload. Files in the directory and its
-    # subdirectories will be uploaded. An empty string means "the current
-    # working directory".
-    # source_directory=""
-
-    # The number of threads to use for the operation. The performance impact of
-    # this value depends on the use case, but generally, smaller files benefit
-    # from more threads and larger files don't benefit from more threads. Too
-    # many threads can slow operations, especially with large files, due to
-    # contention over the Python GIL.
-    # threads=4
-
-    from pathlib import Path
-
-    from google.cloud.storage import Client, transfer_manager
-
-    storage_client = Client()
-    bucket = storage_client.bucket(bucket_name)
-
-    # Generate a list of paths (in string form) relative to the `directory`.
-    # This can be done in a single list comprehension, but is expanded into
-    # multiple lines here for clarity.
-
-    # First, recursively get all files in `directory` as Path objects.
-    directory_as_path_obj = Path(source_directory)
-    paths = directory_as_path_obj.rglob("*")
-
-    # Filter so the list only includes files, not directories themselves.
-    file_paths = [path for path in paths if path.is_file()]
-
-    # These paths are relative to the current working directory. Next, make them
-    # relative to `directory`
-    relative_paths = [path.relative_to(source_directory) for path in file_paths]
-
-    # Finally, convert them all to strings.
-    string_paths = [str(path) for path in relative_paths]
-
-    print("Found {} files.".format(len(string_paths)))
-
-    # Start the upload.
-    results = transfer_manager.upload_many_from_filenames(
-        bucket, string_paths, source_directory=source_directory, threads=threads
-    )
-
-    for name, result in zip(string_paths, results):
-        # The results list is either `None` or an exception for each filename in
-        # the input list, in order.
-
-        if isinstance(result, Exception):
-            print("Failed to upload {} due to exception: {}".format(name, result))
-        else:
-            print("Uploaded {} to {}.".format(name, bucket.name))
-
-
-def download_all_blobs_with_transfer_manager(
-    bucket_name, destination_directory="", threads=4
-):
-    """Download all of the blobs in a bucket, concurrently in a thread pool.
-
-    The filename of each blob once downloaded is derived from the blob name and
-    the `destination_directory `parameter. For complete control of the filename
-    of each blob, use transfer_manager.download_many() instead.
-
-    Directories will be created automatically as needed, for instance to
-    accommodate blob names that include slashes.
-    """
-
-    # The ID of your GCS bucket
-    # bucket_name = "your-bucket-name"
-
-    # The directory on your computer to which to download all of the files. This
-    # string is prepended (with os.path.join()) to the name of each blob to form
-    # the full path. Relative paths and absolute paths are both accepted. An
-    # empty string means "the current working directory". Note that this
-    # parameter allows accepts directory traversal ("../" etc.) and is not
-    # intended for unsanitized end user input.
-    # destination_directory = ""
-
-    # The number of threads to use for the operation. The performance impact of
-    # this value depends on the use case, but generally, smaller files benefit
-    # from more threads and larger files don't benefit from more threads. Too
-    # many threads can slow operations, especially with large files, due to
-    # contention over the Python GIL.
-    # threads=4
-
-    from google.cloud.storage import Client, transfer_manager
-
-    storage_client = Client()
-    bucket = storage_client.bucket(bucket_name)
-
-    blob_names = [blob.name for blob in bucket.list_blobs()]
-
-    results = transfer_manager.download_many_to_path(
-        bucket, blob_names, destination_directory=destination_directory, threads=threads
-    )
-
-    for name, result in zip(blob_names, results):
-        # The results list is either `None` or an exception for each blob in
-        # the input list, in order.
-
-        if isinstance(result, Exception):
-            print("Failed to download {} due to exception: {}".format(name, result))
-        else:
-            print("Downloaded {} to {}.".format(name, destination_directory + name))
